Move urlsForUser function to helper.js file
